feat: Implement calculateArea methods for sub-classes

+++ src/main/java/shapes/Shape.java
@@ -2,5 +2,5 @@ package shapes;
 
 public abstract class Shape {
 
-    abstract int calculateArea();
+    abstract double calculateArea();
 }
